+++ app/services/imap/fetch_email_service.rb
@@ -1,0 +1,15 @@
+class Imap::FetchEmailService < Imap::BaseFetchEmailService
+  def perform
+    fetch_mail_for_channel
+  end
+
+  private
+
+  def authentication_type
+    'PLAIN'
+  end
+
+  def imap_password
+    channel.imap_password
+  end
+end
